api/requirements.txt
# Medical Imaging AI API Requirements
# Phase 5: Backend & API Design

# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.1
opencv-python==4.8.1.78

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlite3  # Built-in with Python
aiosqlite==0.19.0

# Medical Imaging (Optional - for DICOM/NIfTI support)
# pydicom==2.4.3
# nibabel==5.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0

# Monitoring & Logging
structlog==23.2.0
